feat(api): Fetch & Unregister endpoints

+++ internal/repository/auth_repository.go
@@ -16,8 +16,11 @@ import (
 )
 
 type AuthRepoInterface interface {
-	FetchToken(rawToken string) (*tokens.Token, error)
+	TokenExists(rawToken string) (bool, error)
 	CreateToken(request *http.CreateRequest) (string, error)
+	PruneToken(rawToken string) error
+	IsAdmin(rawToken string) (bool, error)
+	FetchToken(rawToken string) (*tokens.Token, error)
 }
 
 type AuthRepository struct {
@@ -64,23 +67,104 @@ func (r *AuthRepository) CreateToken(request *http.CreateRequest) (string, error
 		return "", fmt.Errorf("failed to create user: %w", err)
 	}
 
+	key := r.getCacheKey(utils.Hash(token))
+	err = r.Rdb.Set(ctx, key, "true", r.CacheTTL).Err()
+	if err != nil {
+		return "", err
+	}
+
 	return token, nil
 }
 
-func (r *AuthRepository) FetchToken(rawToken string) (*tokens.Token, error) {
-	key := r.getCacheKey(utils.Hash(rawToken))
+func (r *AuthRepository) IsAdmin(rawToken string) (bool, error) {
+	exists, err := r.TokenExists(rawToken)
+	if err != nil {
+		return false, err
+	}
 
-	query := `SELECT token, admin, upload, delete FROM users WHERE token = $1`
-	var token tokens.Token
-	err := r.SqlClient.QueryRow(ctx, query, utils.Hash(rawToken)).Scan(&token.Data, &token.Permissions.Admin, &token.Permissions.Upload, &token.Permissions.Delete)
+	if !exists {
+		return false, fmt.Errorf("Api token not found")
+	}
+
+	admin := false
+	err = r.SqlClient.QueryRow(
+		ctx,
+		`SELECT admin FROM users WHERE token = $1`,
+		utils.Hash(rawToken),
+	).Scan(&admin)
+
+	if err != nil {
+		return false, err
+	}
+
+	return admin, nil
+}
+
+func (r *AuthRepository) TokenExists(rawToken string) (bool, error) {
+	key := r.getCacheKey(utils.Hash(rawToken))
+	_, err := r.Rdb.Get(ctx, key).Result()
+	if err == nil {
+		return true, nil
+	}
+
+	query := `SELECT 1 FROM users WHERE token = $1`
+	var exists int
+
+	err = r.SqlClient.QueryRow(ctx, query, utils.Hash(rawToken)).Scan(&exists)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, nil
+			return false, nil
 		}
-		return nil, err
+		return false, err
 	}
 
 	err = r.Rdb.Set(ctx, key, "true", r.CacheTTL).Err()
+	if err != nil {
+		return false, err
+	}
+
+	return true, nil
+}
+
+func (r *AuthRepository) PruneToken(rawToken string) error {
+	exists, err := r.TokenExists(rawToken)
+	if err != nil {
+		return err
+	}
+
+	if !exists {
+		return fmt.Errorf("%s does not exists", rawToken)
+	}
+
+	key := r.getCacheKey(utils.Hash(rawToken))
+	_, err = r.Rdb.Del(ctx, key).Result()
+	if err != nil {
+		return err
+	}
+
+	query := `DELETE FROM "users" WHERE "token" = $1`
+	_, err = r.SqlClient.Exec(ctx, query, utils.Hash(rawToken))
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (r *AuthRepository) FetchToken(rawToken string) (*tokens.Token, error) {
+	exists, err := r.TokenExists(rawToken)
+	if err != nil {
+		return nil, err
+	}
+
+	if !exists {
+		return nil, fmt.Errorf("%s api token does not exists", rawToken)
+	}
+
+	query := `SELECT token, admin, upload, delete FROM users WHERE token = $1`
+
+	var token tokens.Token
+	err = r.SqlClient.QueryRow(ctx, query, utils.Hash(rawToken)).Scan(&token.Data, &token.Permissions.Admin, &token.Permissions.Upload, &token.Permissions.Delete)
 	if err != nil {
 		return nil, err
 	}
